+++ sysom-20231230/alibabacloud_sysom20231230/models/_install_agent_with_type_response_body.py
@@ -13,9 +13,15 @@ class InstallAgentWithTypeResponseBody(DaraModel):
         message: str = None,
         request_id: str = None,
     ):
+        # The status code.
+        # - `code == Success` indicates that the authorization is successful.
+        # - Other status codes indicate that the authorization failed. Check the `message` field for the detailed fault information.
         self.code = code
+        # The response data.
         self.data = data
+        # Id of the request
         self.message = message
+        # The request ID, which can be used for end-to-end diagnostics.
         self.request_id = request_id
 
     def validate(self):
@@ -63,6 +69,9 @@ class InstallAgentWithTypeResponseBodyData(DaraModel):
         self,
         task_id: str = None,
     ):
+        # The task ID.
+        # 
+        # You can use this task ID to call the GetAgentTask operation to check the task execution status.
         self.task_id = task_id
 
     def validate(self):