+++ sysom-20231230/alibabacloud_sysom20231230/models/_upgrade_agent_with_type_response_body.py
@@ -13,9 +13,17 @@ class UpgradeAgentWithTypeResponseBody(DaraModel):
         message: str = None,
         request_id: str = None,
     ):
+        # The status code.
+        # - `code == Success` indicates that the authorization is successful.
+        # - Other status codes indicate that the authorization failed. Check the `message` field for the detailed fault information.
         self.code = code
+        # The response data.
         self.data = data
+        # The error message.
+        # - If `code == Success`, this field is empty.
+        # - Otherwise, this field contains the request error information.
         self.message = message
+        # The request ID, which can be used for end-to-end diagnostics.
         self.request_id = request_id
 
     def validate(self):
@@ -63,6 +71,7 @@ class UpgradeAgentWithTypeResponseBodyData(DaraModel):
         self,
         task_id: str = None,
     ):
+        # The task ID.
         self.task_id = task_id
 
     def validate(self):